refactor: move next backoff delay function to retry.rs
